+++ WHATS_NEW.md
@@ -16,6 +16,10 @@ Users
 - My Preferences
 	- Student Fields: search Username
 
+Grades
+
+- Mass Create Assignments
+
 Performance
 
 - 90% gain when updating Side menu